backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
aiosqlite==0.19.0
PyPDF2==3.0.1
python-docx==1.1.0
google-generativeai==0.3.2
